Put the assembly resolver in the package class
Instead of relying on the xaml controls to include the class that hooks
up the resolver, just hook it up on package initialization so it always
runs.
Also, remove useless base package class, it's not really doing anything
useful.
