--- src/GitHub.VisualStudio/UI/Settings/OptionsStyles.xaml
@@ -1,58 +1,0 @@
-﻿<ResourceDictionary
-  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-  xmlns:PresentationOptions="http://schemas.microsoft.com/winfx/2006/xaml/presentation/options">
-  
-  <BorderGapMaskConverter x:Key="BorderGapMaskConverter" />
-  <Style x:Key="GroupBoxFlat" TargetType="{x:Type GroupBox}">
-    <Setter Property="BorderBrush" Value="#D5DFE5" />
-    <Setter Property="BorderThickness" Value="1" />
-    <Setter Property="Template">
-      <Setter.Value>
-        <ControlTemplate TargetType="{x:Type GroupBox}">
-          <Grid SnapsToDevicePixels="true">
-            <Grid.ColumnDefinitions>
-              <ColumnDefinition Width="6" />
-              <ColumnDefinition Width="Auto" />
-              <ColumnDefinition Width="*" />
-              <ColumnDefinition Width="6" />
-            </Grid.ColumnDefinitions>
-            <Grid.RowDefinitions>
-              <RowDefinition Height="Auto" />
-              <RowDefinition Height="Auto" />
-              <RowDefinition Height="*" />
-              <RowDefinition Height="6" />
-            </Grid.RowDefinitions>
-            <Border x:Name="Header"
-                    Grid.Row="0"
-                    Grid.RowSpan="2"
-                    Grid.Column="1"
-                    Padding="3,1,3,0">
-              <ContentPresenter ContentSource="Header" RecognizesAccessKey="True" SnapsToDevicePixels="{TemplateBinding SnapsToDevicePixels}" />
-            </Border>
-            <Border Grid.Row="1"
-                    Grid.RowSpan="3"
-                    Grid.ColumnSpan="4"
-                    BorderBrush="Transparent"
-                    BorderThickness="{TemplateBinding BorderThickness}"
-                    CornerRadius="4">
-              <Border.OpacityMask>
-                <MultiBinding Converter="{StaticResource BorderGapMaskConverter}" ConverterParameter="7">
-                  <Binding ElementName="Header" Path="ActualWidth" />
-                  <Binding Path="ActualWidth" RelativeSource="{RelativeSource Self}" />
-                  <Binding Path="ActualHeight" RelativeSource="{RelativeSource Self}" />
-                </MultiBinding>
-              </Border.OpacityMask>
-              <Border BorderBrush="{TemplateBinding BorderBrush}" BorderThickness="{TemplateBinding BorderThickness}" CornerRadius="3" />
-            </Border>
-            <ContentPresenter Grid.Row="2"
-                              Grid.Column="1"
-                              Grid.ColumnSpan="2"
-                              Margin="{TemplateBinding Padding}"
-                              SnapsToDevicePixels="{TemplateBinding SnapsToDevicePixels}" />
-          </Grid>
-        </ControlTemplate>
-      </Setter.Value>
-    </Setter>
-  </Style>
-</ResourceDictionary>
